+++ src/controllers/llmo.js
@@ -19,10 +19,13 @@ import {
 } from '@adobe/spacecat-shared-utils';
 import { Config } from '@adobe/spacecat-shared-data-access/src/models/site/config.js';
 import crypto from 'crypto';
+import { Entitlement as EntitlementModel } from '@adobe/spacecat-shared-data-access';
+import AccessControlUtil from '../support/access-control-util.js';
 
 const LLMO_SHEETDATA_SOURCE_URL = 'https://main--project-elmo-ui-data--adobe.aem.live';
 
-function LlmoController() {
+function LlmoController(ctx) {
+  const accessControlUtil = AccessControlUtil.fromContext(ctx);
   // Helper function to get site and validate LLMO config
   const getSiteAndValidateLlmo = async (context) => {
     const { siteId } = context.params;
@@ -36,7 +39,9 @@ function LlmoController() {
     if (!llmoConfig?.dataFolder) {
       throw new Error('LLM Optimizer is not enabled for this site, add llmo config to the site');
     }
-
+    if (!await accessControlUtil.hasAccess(site, '', EntitlementModel.PRODUCT_CODES.LLMO)) {
+      throw new Error('Only users belonging to the organization can view its sites');
+    }
     return { site, config, llmoConfig };
   };
 
@@ -110,7 +115,6 @@ function LlmoController() {
       // Get the response data
       const data = await response.json();
 
-      log.info(`Successfully proxied data for siteId: ${siteId}, sheetURL: ${sheetURL}`);
       // Return the data and let the framework handle the compression
       return ok(data, {
         ...(response.headers ? Object.fromEntries(response.headers.entries()) : {}),
@@ -331,6 +335,32 @@ function LlmoController() {
     }
   };
 
+  // Handles requests to the LLMO CDN bucket config endpoint, updates CDN bucket configuration
+  const patchLlmoCdnBucketConfig = async (context) => {
+    const { log } = context;
+    const { data } = context;
+    const { siteId } = context.params;
+
+    try {
+      const { site, config } = await getSiteAndValidateLlmo(context);
+
+      if (!isObject(data)) {
+        return badRequest('Update data must be provided as an object');
+      }
+
+      const { cdnBucketConfig } = data;
+
+      config.updateLlmoCdnBucketConfig(cdnBucketConfig);
+
+      await saveSiteConfig(site, config, log, 'updating CDN logs bucket config');
+
+      return ok(config.getLlmoConfig().cdnBucketConfig || {});
+    } catch (error) {
+      log.error(`Error updating CDN bucket config for siteId: ${siteId}, error: ${error.message}`);
+      return badRequest(error.message);
+    }
+  };
+
   return {
     getLlmoSheetData,
     getLlmoConfig,
@@ -343,6 +373,7 @@ function LlmoController() {
     removeLlmoCustomerIntent,
     patchLlmoCustomerIntent,
     patchLlmoCdnLogsFilter,
+    patchLlmoCdnBucketConfig,
   };
 }