+++ test/controllers/organizations.test.js
@@ -159,6 +159,9 @@ describe('Organizations Controller', () => {
 
     context = {
       dataAccess: mockDataAccess,
+      pathInfo: {
+        headers: { 'x-product': 'abcd' },
+      },
       attributes: {
         authInfo: new AuthInfo()
           .withType('jwt')