+++ test/index.test.js
@@ -50,6 +50,18 @@ describe('Index Tests', () => {
     }),
   };
 
+  const scrapeJobConfiguration = {
+    queues: ['spacecat-scrape-queue-1', 'spacecat-scrape-queue-2', 'spacecat-scrape-queue-3'],
+    scrapeWorkerQueue: 'https://sqs.us-east-1.amazonaws.com/1234567890/scrape-worker-queue',
+    scrapeQueueUrlPrefix: 'https://sqs.us-east-1.amazonaws.com/1234567890/',
+    s3Bucket: 's3-bucket',
+    options: {
+      enableJavascript: true,
+      hideConsentBanners: false,
+    },
+    maxUrlsPerJob: 3,
+  };
+
   beforeEach('setup', () => {
     // Mock AccessControlUtil's hasAdminAccess method
     sinon.stub(AccessControlUtil.prototype, 'hasAdminAccess').returns(true);
@@ -76,6 +88,10 @@ describe('Index Tests', () => {
         IMS_CLIENT_CODE: 'mock-client-code',
         IMS_CLIENT_SECRET: 'mock-client-secret',
         IMPORT_CONFIGURATION: '{}',
+        REPORT_JOBS_QUEUE_URL: 'https://sqs.example.com/reports-queue',
+        S3_REPORT_BUCKET: 'test-reports-bucket',
+        S3_MYSTIQUE_BUCKET: 'test-mystique-bucket',
+        SCRAPE_JOB_CONFIGURATION: JSON.stringify(scrapeJobConfiguration),
       },
       dataAccess: {
         Audit: {